rag/requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
faiss-cpu>=1.7.4
tiktoken>=0.5.2
numpy>=1.24.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0